// src/api/__generated__/story/model/actorCollectionInfoResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

/**
 * 关联演员 IP 列表（computingPower 即片酬）
 */
export interface ActorCollectionInfoResponse {
  actorCollectionId?: number;
  actorCollectionName?: string;
  actorCollectionAvatar?: string;
  badge?: string;
  trust?: number;
  computingPower?: number;
}