// src/api/__generated__/story/model/dramaDetailResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { BoundActorCollectionResponse } from './boundActorCollectionResponse';
import type { DramaDetailResponseBadge } from './dramaDetailResponseBadge';

export interface DramaDetailResponse {
  id?: number;
  userId?: number;
  title?: string;
  description?: string;
  coverUrl?: string;
  totalEpisodes?: number;
  tags?: string[];
  boundActorCollections?: BoundActorCollectionResponse[];
  status?: string;
  badge?: DramaDetailResponseBadge;
  source?: string;
  auditReason?: string;
  onlineAt?: number;
  offlineAt?: number;
  nftChain?: string;
  nftTokenStandard?: string;
  nftContractAddress?: string;
  nftTxHash?: string;
  nftMinted?: boolean;
  createdAt?: number;
  updatedAt?: number;
  version?: number;
}