// src/api/__generated__/story/model/listMyShortVideosParams.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { ListMyShortVideosStatus } from './listMyShortVideosStatus';

export type ListMyShortVideosParams = {
  /**
   * 短视频状态
   */
  status?: ListMyShortVideosStatus;
  /**
   * @minimum -1
   */
  mark?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};