// src/api/__generated__/story/model/listMyShortVideosStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export type ListMyShortVideosStatus =
  (typeof ListMyShortVideosStatus)[keyof typeof ListMyShortVideosStatus];

export const ListMyShortVideosStatus = {
  PENDING_REVIEW: 'PENDING_REVIEW',
  REVIEW_REJECTED: 'REVIEW_REJECTED',
  ONLINE: 'ONLINE',
  OFFLINE: 'OFFLINE',
  DELETED: 'DELETED',
} as const;